+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.0
 * feat: Rename `list` commands to `ls`.
+* feat: Add `lambda-ai instance-types --available-only` to only list available instance types.
 
 # 1.0.0
 * feat: Initial release.